requirements.txt
discord.py>=2.0.0
google-generativeai>=0.8.0
python-dotenv>=1.0.0
requests>=2.28.0
beautifulsoup4>=4.12.0
